Creation of function to compute ∆gi components
Creation of a function to calculate the ∆g x/y/z components. Currently this is a separate function to the final compute gz function, but depending on complexity they could be merged. Additional notes & questions section added to the end of notebook to keep track of issues. This was also included in issue #7.