Changes to structure, addition of prolate function
Change of name as, for now, all code will be held here until a clearer idea of how to separate it is formed. A function for the prolate ellipsoid case was added. This commit also includes some additional comments for clarity, and some raise error additions to ensure each function is used with the correct relationship between the semi axes values (e.g. a > b > c).