Bug fix in calc_lambda, addition of get_mask, breakdown of get_ellipsoids
Adding np.any() so lambda function works for arrays. Addition of function to get the mask for internal/external points. Addition of get_ABC (in get_gravity...) as A(lambda) etc equations are used in both trixaial and in internal functions. Update of docstrings for clarity. Update of __init___.py.