Fix bug in computation of the magnetization vector
Fix bug in the `_get_magnetisation` function related to a wrong sign
when computing the magnetization vector of an ellipsoid considering the
demagnetization effect. Add tests for checking that the magnetization
considering demagnetization effect is smaller than the magnetization
without considering it, and to check that the components of the
internal demagnetization tensor are all positive.